[ Feat ] - Criando rota vendedor, desenvolvendo novos dominios read necessarios para regras de negocio

+++ src/main/java/com/logimarui/core/api/domain/read/Vendedor.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class Vendedor {
     private Long codigo;
     private String nome;
-    private Long area;
+    private Integer area;
 }
